Abstracting out spreads

+++ src/main/java/org/kutsuki/zerotwo/portfolio/PortfolioManager.java
@@ -1,17 +1,15 @@
 package org.kutsuki.zerotwo.portfolio;
 
-import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.time.format.DateTimeFormatter;
-import java.time.format.DateTimeFormatterBuilder;
-import java.time.format.DateTimeParseException;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashMap;
+import java.util.Iterator;
 import java.util.List;
-import java.util.Locale;
 import java.util.Map;
+import java.util.Set;
 
 import javax.annotation.PostConstruct;
 
@@ -25,32 +23,23 @@ import org.springframework.stereotype.Component;
 
 @Component
 public class PortfolioManager {
-    private static final DateTimeFormatter FORMATTER = new DateTimeFormatterBuilder().parseCaseInsensitive()
-	    .appendPattern("d MMM yy").toFormatter(Locale.ENGLISH);
     private static final DateTimeFormatter TIME_FORMATTER = DateTimeFormatter.ofPattern("h:mma");
-    private static final String BACKRATIO = "BACKRATIO";
     private static final String BOLD = "<b>";
     private static final String BOLD_CLOSE = "</b>";
     private static final String BOT = "BOT ";
     private static final String BUY = "BUY ";
-    private static final String BUTTERFLY = "BUTTERFLY";
-    private static final String CONDOR = "CONDOR";
-    private static final String HUNDRED = "100";
-    private static final String IRON_CONDOR = "IRON CONDOR";
     private static final String NEW = "NEW";
     private static final String PORTFOLIO = "[PORTFOLIO]";
     private static final String SELL = "SELL ";
     private static final String SOLD = "SOLD ";
     private static final String ST_WEEKLY = "ST Weekly";
-    private static final String WEEKLYS = "(Weeklys)";
     private static final String WORKING = "WORKING";
     private static final String WORKING_EXPLAINATION = " - This trade has not been filled!";
-    private static final String VERTICAL = "VERTICAL";
-    private static final String UNBALANCED_BUTTERFLY = "~BUTTERFLY";
     private static final String UNOFFICIAL = "Unofficial:";
 
     private List<Position> deleteList;
     private List<Position> saveList;
+    private List<AbstractSpread> spreadList;
     private Map<String, Position> portfolioMap;
 
     @Autowired
@@ -67,6 +56,17 @@ public class PortfolioManager {
 	this.deleteList = new ArrayList<Position>();
 	this.saveList = new ArrayList<Position>();
 	this.portfolioMap = new HashMap<String, Position>();
+
+	// order matters
+	this.spreadList = new ArrayList<AbstractSpread>();
+	this.spreadList.add(new IronCondorSpread());
+	this.spreadList.add(new UnbalancedButterflySpread());
+	this.spreadList.add(new RatioSpread());
+	this.spreadList.add(new VerticalSpread());
+	this.spreadList.add(new ButterflySpread());
+	this.spreadList.add(new CondorSpread());
+	this.spreadList.add(new SingleSpread());
+
 	reloadCache();
     }
 
@@ -174,11 +174,20 @@ public class PortfolioManager {
 	    service.email(emailPortfolio, subject.toString(), body.toString());
 
 	    // update portfolio
-	    repository.deleteAll(deleteList);
-	    repository.saveAll(saveList);
+	    if (!deleteList.isEmpty()) {
+		repository.deleteAll(deleteList);
+	    }
+
+	    if (!saveList.isEmpty()) {
+		repository.saveAll(saveList);
+	    }
 	} catch (Exception e) {
 	    service.emailException(escaped, e);
 	}
+    }
+
+    public Set<String> getSymbols() {
+	return portfolioMap.keySet();
     }
 
     public void reloadCache() {
@@ -188,8 +197,8 @@ public class PortfolioManager {
 	}
     }
 
-    public List<String> updateQty(String symbol, String qty) {
-	List<String> result = new ArrayList<String>();
+    public String updateQty(String symbol, String qty) {
+	String result = null;
 
 	if (StringUtils.isNotBlank(symbol)) {
 	    Position position = portfolioMap.get(symbol);
@@ -203,9 +212,7 @@ public class PortfolioManager {
 		}
 	    }
 
-	    result.add(position.toString());
-	} else {
-	    result.addAll(portfolioMap.keySet());
+	    result = position.toString();
 	}
 
 	return result;
@@ -229,20 +236,15 @@ public class PortfolioManager {
 		    }
 
 		    String[] split2 = StringUtils.split(split, StringUtils.SPACE);
-		    if (StringUtils.contains(split, BACKRATIO)) {
-			orderList.add(backratio(split2));
-		    } else if (StringUtils.contains(split, IRON_CONDOR)) {
-			orderList.add(ironCondor(split2));
-		    } else if (StringUtils.contains(split, UNBALANCED_BUTTERFLY)) {
-			orderList.add(unbalancedButterfly(split2));
-		    } else if (StringUtils.contains(split, VERTICAL)) {
-			orderList.add(vertical(split2));
-		    } else if (StringUtils.contains(split, BUTTERFLY)) {
-			orderList.add(butterfly(split2));
-		    } else if (StringUtils.contains(split, CONDOR)) {
-			orderList.add(condor(split2));
-		    } else {
-			orderList.add(single(split2));
+		    boolean found = false;
+		    Iterator<AbstractSpread> itr = spreadList.iterator();
+		    while (!found && itr.hasNext()) {
+			AbstractSpread spread = itr.next();
+
+			if (!itr.hasNext() || StringUtils.contains(split, spread.getSpread())) {
+			    orderList.add(spread.parseOrder(split2));
+			    found = true;
+			}
 		    }
 		}
 	    } catch (Exception e) {
@@ -318,277 +320,5 @@ public class PortfolioManager {
 	}
 
 	return sb.toString();
-    }
-
-    private OrderModel backratio(String[] split) throws Exception {
-	int quantity = parseQuantity(split[0]);
-	List<BigDecimal> ratioList = parseSlashes(split[1]);
-	String symbol = parseSymbol(split[3]);
-
-	int i = 0;
-	if (StringUtils.equalsIgnoreCase(split[4], HUNDRED)) {
-	    i++;
-	}
-
-	if (StringUtils.equalsIgnoreCase(split[5], WEEKLYS)) {
-	    i++;
-	}
-
-	LocalDate expiry = parseExpiry(split[4 + i], split[5 + i], split[6 + i]);
-	List<BigDecimal> strikeList = parseSlashes(split[7 + i]);
-	OptionType type = parseType(split[8 + i]);
-	BigDecimal price = parsePrice(split[9 + i]);
-
-	OrderModel order = new OrderModel(BACKRATIO, price, split[9 + i]);
-	int qty1 = -quantity * ratioList.get(0).intValue();
-	order.addPosition(new Position(qty1, symbol, expiry, strikeList.get(0), type));
-	int qty2 = quantity * ratioList.get(1).intValue();
-	order.addPosition(new Position(qty2, symbol, expiry, strikeList.get(1), type));
-
-	return order;
-    }
-
-    private OrderModel butterfly(String[] split) throws Exception {
-	int quantity = parseQuantity(split[0]);
-	String symbol = parseSymbol(split[2]);
-
-	int i = 0;
-	if (StringUtils.equalsIgnoreCase(split[3], HUNDRED)) {
-	    i++;
-	}
-
-	if (StringUtils.equalsIgnoreCase(split[4], WEEKLYS)) {
-	    i++;
-	}
-
-	LocalDate expiry = parseExpiry(split[3 + i], split[4 + i], split[5 + i]);
-	List<BigDecimal> strikeList = parseSlashes(split[6 + i]);
-	OptionType type = parseType(split[7 + i]);
-	BigDecimal price = parsePrice(split[8 + i]);
-
-	OrderModel order = new OrderModel(BUTTERFLY, price, split[8 + i]);
-	order.addPosition(new Position(quantity, symbol, expiry, strikeList.get(0), type));
-	int qty2 = -quantity * 2;
-	order.addPosition(new Position(qty2, symbol, expiry, strikeList.get(1), type));
-	order.addPosition(new Position(quantity, symbol, expiry, strikeList.get(2), type));
-
-	return order;
-    }
-
-    private OrderModel condor(String[] split) throws Exception {
-	int quantity = parseQuantity(split[0]);
-	String symbol = parseSymbol(split[2]);
-
-	int i = 0;
-	if (StringUtils.equalsIgnoreCase(split[3], HUNDRED)) {
-	    i++;
-	}
-
-	if (StringUtils.equalsIgnoreCase(split[4], WEEKLYS)) {
-	    i++;
-	}
-
-	LocalDate expiry = parseExpiry(split[3 + i], split[4 + i], split[5 + i]);
-	List<BigDecimal> strikeList = parseSlashes(split[6 + i]);
-	OptionType type = parseType(split[7 + i]);
-	BigDecimal price = parsePrice(split[8 + i]);
-
-	OrderModel order = new OrderModel(CONDOR, price, split[8 + i]);
-	order.addPosition(new Position(quantity, symbol, expiry, strikeList.get(0), type));
-	order.addPosition(new Position(-quantity, symbol, expiry, strikeList.get(1), type));
-	order.addPosition(new Position(-quantity, symbol, expiry, strikeList.get(2), type));
-	order.addPosition(new Position(quantity, symbol, expiry, strikeList.get(3), type));
-
-	return order;
-    }
-
-    private OrderModel ironCondor(String[] split) throws Exception {
-	int quantity = parseQuantity(split[0]);
-	String symbol = parseSymbol(split[3]);
-
-	int i = 0;
-	if (StringUtils.equalsIgnoreCase(split[4], HUNDRED)) {
-	    i++;
-	}
-
-	if (StringUtils.equalsIgnoreCase(split[5], WEEKLYS)) {
-	    i++;
-	}
-
-	LocalDate expiry = parseExpiry(split[4 + i], split[5 + i], split[6 + i]);
-	List<BigDecimal> strikeList = parseSlashes(split[7 + i]);
-	List<OptionType> type = parseTypes(split[8 + i]);
-	BigDecimal price = parsePrice(split[9 + i]);
-
-	OrderModel order = new OrderModel(IRON_CONDOR, price, split[9 + i]);
-	order.addPosition(new Position(quantity, symbol, expiry, strikeList.get(0), type.get(0)));
-	order.addPosition(new Position(-quantity, symbol, expiry, strikeList.get(1), type.get(0)));
-	order.addPosition(new Position(quantity, symbol, expiry, strikeList.get(2), type.get(1)));
-	order.addPosition(new Position(quantity, symbol, expiry, strikeList.get(3), type.get(1)));
-
-	return order;
-    }
-
-    private OrderModel single(String[] split) throws Exception {
-	int quantity = parseQuantity(split[0]);
-	String symbol = parseSymbol(split[1]);
-
-	int i = 0;
-	if (StringUtils.equalsIgnoreCase(split[2], HUNDRED)) {
-	    i++;
-	}
-
-	if (StringUtils.equalsIgnoreCase(split[3], WEEKLYS)) {
-	    i++;
-	}
-
-	LocalDate expiry = parseExpiry(split[2 + i], split[3 + i], split[4 + i]);
-	BigDecimal strike = parseStrike(split[5 + i]);
-	OptionType type = parseType(split[6 + i]);
-	BigDecimal price = parsePrice(split[7 + i]);
-
-	OrderModel order = new OrderModel(type.toString(), price, split[7 + i]);
-	order.addPosition(new Position(quantity, symbol, expiry, strike, type));
-
-	return order;
-    }
-
-    private OrderModel unbalancedButterfly(String[] split) throws Exception {
-	int quantity = parseQuantity(split[0]);
-	List<BigDecimal> ratioList = parseSlashes(split[1]);
-	String symbol = parseSymbol(split[3]);
-
-	int i = 0;
-	if (StringUtils.equalsIgnoreCase(split[4], HUNDRED)) {
-	    i++;
-	}
-
-	if (StringUtils.equalsIgnoreCase(split[5], WEEKLYS)) {
-	    i++;
-	}
-
-	LocalDate expiry = parseExpiry(split[4 + i], split[5 + i], split[6 + i]);
-	List<BigDecimal> strikeList = parseSlashes(split[7 + i]);
-	OptionType type = parseType(split[8 + i]);
-	BigDecimal price = parsePrice(split[9 + i]);
-
-	OrderModel order = new OrderModel(UNBALANCED_BUTTERFLY, price, split[9 + i]);
-	int qty1 = quantity * ratioList.get(0).intValue();
-	order.addPosition(new Position(qty1, symbol, expiry, strikeList.get(0), type));
-	int qty2 = -quantity * ratioList.get(1).intValue();
-	order.addPosition(new Position(qty2, symbol, expiry, strikeList.get(1), type));
-	int qty3 = quantity * ratioList.get(2).intValue();
-	order.addPosition(new Position(qty3, symbol, expiry, strikeList.get(2), type));
-
-	return order;
-    }
-
-    private OrderModel vertical(String[] split) throws Exception {
-	int quantity = parseQuantity(split[0]);
-	String symbol = parseSymbol(split[2]);
-
-	int i = 0;
-	if (StringUtils.equalsIgnoreCase(split[3], HUNDRED)) {
-	    i++;
-	}
-
-	if (StringUtils.equalsIgnoreCase(split[4], WEEKLYS)) {
-	    i++;
-	}
-
-	LocalDate expiry = parseExpiry(split[3 + i], split[4 + i], split[5 + i]);
-	List<BigDecimal> strikeList = parseSlashes(split[6 + i]);
-	OptionType type = parseType(split[7 + i]);
-	BigDecimal price = parsePrice(split[8 + i]);
-
-	OrderModel order = new OrderModel(VERTICAL, price, split[8 + i]);
-	order.addPosition(new Position(quantity, symbol, expiry, strikeList.get(0), type));
-	order.addPosition(new Position(-quantity, symbol, expiry, strikeList.get(1), type));
-
-	return order;
-    }
-
-    private LocalDate parseExpiry(String day, String month, String year) throws Exception {
-	LocalDate exp = null;
-
-	try {
-	    String date = day + StringUtils.SPACE + month + StringUtils.SPACE + year;
-	    exp = LocalDate.parse(date, FORMATTER);
-	} catch (DateTimeParseException e) {
-	    throw new Exception("Error parsing Date: " + day + StringUtils.SPACE + month + StringUtils.SPACE + year, e);
-	}
-
-	return exp;
-    }
-
-    private BigDecimal parsePrice(String val) throws Exception {
-	BigDecimal price = null;
-
-	if (!StringUtils.contains(val, '@') || !StringUtils.contains(val, '.')) {
-	    throw new Exception("Missing @ or . from price: " + val);
-	}
-
-	try {
-	    price = new BigDecimal(StringUtils.substring(val, 1, StringUtils.indexOf(val, '.') + 3));
-	} catch (NumberFormatException e) {
-	    throw new Exception("Error parsing price: " + val, e);
-	}
-
-	return price;
-    }
-
-    private int parseQuantity(String val) throws Exception {
-	int qty = 0;
-
-	try {
-	    qty = Integer.parseInt(val);
-	} catch (NumberFormatException e) {
-	    throw new Exception("Error parsing quantity: " + val, e);
-	}
-
-	return qty;
-    }
-
-    private List<BigDecimal> parseSlashes(String slashes) throws Exception {
-	List<BigDecimal> slashList = new ArrayList<BigDecimal>();
-
-	for (String slash : StringUtils.split(slashes, '/')) {
-	    try {
-		slashList.add(new BigDecimal(slash));
-	    } catch (NumberFormatException e) {
-		throw new Exception("Error parsing slashes: " + slashes, e);
-	    }
-	}
-
-	return slashList;
-    }
-
-    private BigDecimal parseStrike(String strike) throws Exception {
-	return parseSlashes(strike).get(0);
-    }
-
-    private String parseSymbol(String symbol) throws Exception {
-	if (StringUtils.length(symbol) > 4 || StringUtils.length(symbol) == 0) {
-	    throw new Exception("Error parsing symbol: " + symbol);
-	}
-
-	return symbol;
-    }
-
-    private OptionType parseType(String type) throws Exception {
-	return parseTypes(type).get(0);
-    }
-
-    private List<OptionType> parseTypes(String slashes) throws Exception {
-	List<OptionType> slashList = new ArrayList<OptionType>();
-	for (String split : StringUtils.split(slashes, '/')) {
-	    try {
-		slashList.add(OptionType.valueOf(split));
-	    } catch (IllegalArgumentException e) {
-		throw new Exception("Error parsing type: " + split);
-	    }
-	}
-
-	return slashList;
     }
 }
